add external accessable contents to __init__.py
This might be due to my bad understanding of python packaging, but
without I am unable to access the contents of the library on my system.

Previously, accessing from the directory works. Run 'pip install -e ...'
and access from some other directory. On my system the package
'PyOrgMode' is empty.